refactor(web): extract notifications list for cleaner state management
