+++ api/src/unraid-api/types/util.ts
